simple_dns::rdata

Struct SVCB

pub struct SVCB<'a> {
    pub priority: u16,
    pub target: Name<'a>,
    /* private fields */
}

The SVCB DNS RR type is used to locate alternative endpoints for a service. RFC 9460.

Fields
priority: u16

The priority of this record (relative to others, with lower values preferred).

A value of 0 indicates AliasMode.

target: Name<'a>

The domain name of either the alias target (for AliasMode) or the alternative endpoint (for ServiceMode).

Implementations

impl<'a> SVCB<'a>

pub fn new(priority: u16, target: Name<'a>) -> Self

Creates a new SVCB instance with no parameters.

pub fn set_param(&mut self, param: SVCParam<'a>)

Sets a parameter, replacing any previous value.

pub fn with_param(self, param: SVCParam<'a>) -> Self

Same as Self::set_param, but returns self for chaining.

pub fn set_mandatory(&mut self, keys: impl Iterator<Item = u16>)

Sets the โmandatoryโ parameter.

If keys is empty, this method does nothing.

pub fn set_alpn(&mut self, alpn_ids: &[CharacterString<'a>])

Sets the โalpnโ parameter.

if alpn_ids is empty, this method does nothing.

pub fn set_no_default_alpn(&mut self)

Sets the โno-default-alpnโ parameter.

pub fn set_port(&mut self, port: u16)

Sets the โportโ parameter.

pub fn set_ipv4hint(&mut self, ips: &[u32])

Sets the โipv4hintโ parameter.

if ips is empty, this method does nothing.

pub fn set_ipv6hint(&mut self, ips: &[u128])

Sets the โipv6hintโ parameter.

if ips is empty, this method does nothing

pub fn get_param(&'a self, key: u16) -> Option<&'a SVCParam<'a>>

Gets a read-only reference to the SVCParam

Returns None if the key does not exist.

pub fn iter_params(&self) -> impl Iterator<Item = &SVCParam<'a>>

Iterates over all parameters.

pub fn into_owned<'b>(self) -> SVCB<'b>

Transforms the inner data into its owned type