Struct simple_dns::Name

pub struct Name<'a> { /* private fields */ }

A Name represents a domain-name, which consists of character strings separated by dots.
Each section of a name is called label
ex: google.com consists of two labels google and com

Implementations

impl<'a> Name<'a>

pub fn new(name: &'a str) -> Result<Self>

Creates a new validated Name

pub fn new_unchecked(name: &'a str) -> Self

Create a new Name without checking for size limits

pub fn is_link_local(&self) -> bool

Verify if name ends with .local.

pub fn iter(&'a self) -> Iter<'_, Label<'a>>

Returns an Iter of this Name Labels

pub fn is_subdomain_of(&self, other: &Name<'_>) -> bool

Returns true if self is a subdomain of other

pub fn without(&self, domain: &Name<'_>) -> Option<Name<'_>>

Returns the subdomain part of self, based on domain. If self is not a subdomain of domain, returns None

Example:

let name = Name::new_unchecked("sub.domain.local");
let domain = Name::new_unchecked("domain.local");

assert!(domain.without(&name).is_none());

let sub = name.without(&domain).unwrap();
assert_eq!(sub.to_string(), "sub")

pub fn into_owned<'b>(self) -> Name<'b>

Transforms the inner data into its owned type

pub fn get_labels(&self) -> &[Label<'_>]

Get the labels that compose this name