simple_dns

Struct Packet

pub struct Packet<'a> {
    pub questions: Vec<Question<'a>>,
    pub answers: Vec<ResourceRecord<'a>>,
    pub name_servers: Vec<ResourceRecord<'a>>,
    pub additional_records: Vec<ResourceRecord<'a>>,
    /* private fields */
}

Represents a DNS message packet

When working with EDNS packets, use Packet::opt and Packet::opt_mut to add or access OPT packet information

Fields
questions: Vec<Question<'a>>

Questions section

answers: Vec<ResourceRecord<'a>>

Answers section

name_servers: Vec<ResourceRecord<'a>>

Name servers section

additional_records: Vec<ResourceRecord<'a>>

Aditional records section.
DO NOT use this field to add OPT record, use [opt_mut] instead

Implementations

impl<'a> Packet<'a>

pub fn new_query(id: u16) -> Self

Creates a new empty packet with a query header

pub fn new_reply(id: u16) -> Self

Creates a new empty packet with a reply header

pub fn id(&self) -> u16

Get packet id

pub fn set_flags(&mut self, flags: PacketFlag)

Set flags in the packet

pub fn remove_flags(&mut self, flags: PacketFlag)

Remove flags present in the packet

pub fn has_flags(&self, flags: PacketFlag) -> bool

Check if the packet has flags set

pub fn rcode(&self) -> RCODE

Get this packet RCODE information

pub fn rcode_mut(&mut self) -> &mut RCODE

Get a mutable reference for this packet RCODE information Warning, if the RCODE value is greater than 15 (4 bits), you MUST provide an OPT resource record through the [opt_mut] function

pub fn opcode(&self) -> OPCODE

Get this packet OPCODE information

pub fn opcode_mut(&mut self) -> &mut OPCODE

Get a mutable reference for this packet OPCODE information

pub fn opt(&self) -> Option<&OPT<'a>>

Get the OPT resource record for this packet, if present

pub fn opt_mut(&mut self) -> &mut Option<OPT<'a>>

Get a mutable reference for this packet OPT resource record.

pub fn into_reply(self) -> Self

Changes this packet into a reply packet by replacing its header

pub fn parse(data: &'a [u8]) -> Result<Self>

Parses a packet from a slice of bytes

pub fn build_bytes_vec(&self) -> Result<Vec<u8>>

Creates a new Vec<u8> and write the contents of this package in wire format

This call will allocate a Vec<u8> of 900 bytes, which is enough for a jumbo UDP packet

pub fn build_bytes_vec_compressed(&self) -> Result<Vec<u8>>

Creates a new Vec<u8> and write the contents of this package in wire format with compression enabled

This call will allocate a Vec<u8> of 900 bytes, which is enough for a jumbo UDP packet

pub fn write_to<T: Write>(&self, out: &mut T) -> Result<()>

Write the contents of this package in wire format into the provided writer

pub fn write_compressed_to<T: Write + Seek>(&self, out: &mut T) -> Result<()>

Write the contents of this package in wire format with enabled compression into the provided writer