simple_dns

Enum RCODE

pub enum RCODE {

    NoError = 0,
    FormatError = 1,
    ServerFailure = 2,
    NameError = 3,
    NotImplemented = 4,
    Refused = 5,
    YXDOMAIN = 6,
    YXRRSET = 7,
    NXRRSET = 8,
    NOTAUTH = 9,
    NOTZONE = 10,
    BADVERS = 16,
    Reserved = 17,

}

Possible RCODE values for a DNS packet
RFC 1035 Response code - this 4 bit field is set as part of responses.
The values have the following interpretation

Variants

NoError = 0

No error condition

FormatError = 1

Format error - The name server was unable to interpret the query.

ServerFailure = 2

Server failure - The name server was unable to process this query due to a problem with the name server.

NameError = 3

Name Error - Meaningful only for responses from an authoritative name server,
this code signifies that the domain name referenced in the query does not exist.

NotImplemented = 4

Not Implemented - The name server does not support the requested kind of query.

Refused = 5

Refused - The name server refuses to perform the specified operation for policy reasons.
For example, a name server may not wish to provide the information to the particular requester,
or a name server may not wish to perform a particular operation (e.g., zone transfer) for particular data.

YXDOMAIN = 6

Some name that ought not to exist, does exist. RFC 2136

YXRRSET = 7

Some RRset that ought not to exist, does exist. RFC 2136

NXRRSET = 8

Some RRset that ought to exist, does not exist. RFC 2136

NOTAUTH = 9

The server is not authoritative for the zone named in the Zone Section. RFC 2136

NOTZONE = 10

A name used in the Prerequisite or Update Section is not within the zone denoted by the Zone Section. RFC 2136

BADVERS = 16

EDNS Version not supported by the responder RFC 6891

Reserved = 17

Reserved for future use.