simple_dns

Module rdata

Contains RData implementations

Structs

A

Represents a Resource Address (IPv4)

AAAA

Represents a Resource Address (IPv6) rfc3596

AFSDB

AFSDB records represents servers with ASD cells

CAA

RFC 8659: Allow domain name holders to indicate whether they are authorized to issue digital certificates for particular domain name Used as a security policy for certificate authorities This implementation does not validate the tag or value; it splits based on packet byte structure

CNAME

Canonical name for an alias, RFC 1035

HINFO

HINFO records are used to acquire general information about a host.
The main use is for protocols such as FTP that can use special procedures when talking between machines or operating systems of the same type.

ISDN

An ISDN (Integrated Service Digital Network) number is simply a telephone number.

LOC

A Means for Expressing Location Information in the Domain Name System RFC 1876

MB

Mailbox domain name (EXPERIMENTAL), RFC 1035

MD

Mail destination (Obsolete - use MX), RFC 1035

MF

Mail forwarder (Obsolete - use MX), RFC 1035

MG

Mail group member (EXPERIMENTAL), RFC 1035

MINFO

MINFO recors are used to acquire mailbox or mail list information

MR

Mail rename domain name (EXPERIMENTAL), RFC 1035

MX

MX is used to acquire mail exchange information

NS

Authoritative name server, RFC 1035

NSAP

NSAP structure RFC 1706
ATTENTION: this code doesn’t validade the content of the NSAP RR, it just split the bytes in the correct order

NSAP_PTR

PTR for NSAP records, RFC 1348

NULL

NULL resources are used to represent any kind of information.

OPT

OPT is a pseudo-rr used to carry control information
If an OPT record is present in a received request, responders MUST include an OPT record in their respective responses.
OPT RRs MUST NOT be cached, forwarded, or stored in or loaded from master files.

OPTCode

Represents the variable part of an OPT rr

PTR

Domain name pointer, RFC 1035

RP

RP Responsible Person, RFC 1183

RouteThrough

The RT resource record provides a route-through binding for hosts that do not have their own direct wide area network addresses

SOA

SOA records are used to mark the start of a zone of authority

SRV

SRV records specifies the location of the server(s) for a specific protocol and domain.

TXT

Represents a TXT Resource Record

WKS

The WKS record is used to describe the well known services supported by a particular protocol on a particular internet address.

X25

X.25 address, RFC 1183

Enums

RData

Represents the RData of each TYPE

TYPE

Possible TYPE values in DNS Resource Records Each value is described according to its own RFC